# backend/docs/openapi-style.yaml
openapi: 3.1.0
info:
  title: Channels Backend API
  version: 0.1.0
  description: OpenAPI-style documentation generated from the current backend implementation.
servers:
  - url: http://localhost:8000
tags:
  - name: auth
  - name: channels
  - name: memberships
  - name: messages
  - name: health
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  username:
                    type: string
                  email:
                    type: string
                    format: email
                    nullable: true
                required: [id, username]
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /auth/login:
    post:
      tags: [auth]
      summary: Login with username/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke a refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required: [status]
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /channels:
    post:
      tags: [channels]
      summary: Create a channel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelCreateRequest'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [channels]
      summary: List visible channels for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Channel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}:
    get:
      tags: [channels]
      summary: Get channel by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Channel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/join:
    post:
      tags: [memberships]
      summary: Join a channel (behavior depends on join mode)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRequest'
      responses:
        '200':
          description: Membership state after join
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipActionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /channels/{channel_id}/invite:
    post:
      tags: [memberships]
      summary: Create channel invite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteRequest'
      responses:
        '200':
          description: Invite generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /invites/{token}/accept:
    post:
      tags: [memberships]
      summary: Accept invite by token
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Membership state after invite acceptance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipActionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/members/{user_id}/approve:
    post:
      tags: [memberships]
      summary: Approve pending member
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Approved membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipActionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/members/{user_id}/add:
    post:
      tags: [memberships]
      summary: Add user directly as member
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Membership after add
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipActionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/members/{user_id}/promote:
    post:
      tags: [memberships]
      summary: Promote member to admin
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Membership after promote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipActionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/members/{user_id}/demote:
    post:
      tags: [memberships]
      summary: Demote admin to member
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Membership after demote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipActionResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/members/{user_id}:
    delete:
      tags: [memberships]
      summary: Remove member from channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Removal successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required: [status]
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/messages:
    post:
      tags: [messages]
      summary: Publish message to channel
      description: Only owner/admin can publish.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishMessageRequest'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [messages]
      summary: List channel messages
      description: Requires read permission (owner/admin/member).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - name: before_seq_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Messages in descending seq order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'

  /channels/{channel_id}/seen:
    post:
      tags: [messages]
      summary: Update seen/read state
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeenRequest'
      responses:
        '200':
          description: Seen state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeenResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /health:
    get:
      tags: [health]
      summary: Liveness/readiness style health check
      responses:
        '200':
          description: Service dependencies are reachable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  db:
                    type: boolean
                  redis:
                    type: boolean
                  rabbitmq:
                    type: boolean
                required: [status, db, redis, rabbitmq]

  /ws:
    get:
      tags: [messages]
      summary: WebSocket endpoint
      description: >
        Not a standard REST endpoint. Connect using WebSocket with access token in query `token`
        or `Authorization: Bearer <token>` header.
      x-websocket:
        auth:
          query_token: true
          bearer_header: true
        server_messages:
          - type: hello
          - type: history
          - type: message
        client_messages:
          - type: sync
          - type: seen
      responses:
        '101':
          description: Switching Protocols

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ChannelId:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    ErrorResponse:
      description: Application-level error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
            required: [detail]
    ValidationError:
      description: Request validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
            required: [detail]

  schemas:
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
          minLength: 8
          maxLength: 256
      required: [username, password]

    LoginRequest:
      type: object
      properties:
        username_or_email:
          type: string
        password:
          type: string
      required: [username_or_email, password]

    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]

    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]

    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
      required: [access_token, refresh_token, token_type]

    ChannelVisibility:
      type: string
      enum: [public, private]

    ChannelJoinMode:
      type: string
      enum: [open, invite_only, approval_required]

    MembershipRole:
      type: string
      enum: [owner, admin, member, pending]

    ChannelCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        visibility:
          $ref: '#/components/schemas/ChannelVisibility'
        join_mode:
          $ref: '#/components/schemas/ChannelJoinMode'
      required: [name, visibility, join_mode]

    ChannelResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_user_id:
          type: string
          format: uuid
        name:
          type: string
        visibility:
          $ref: '#/components/schemas/ChannelVisibility'
        join_mode:
          $ref: '#/components/schemas/ChannelJoinMode'
        created_at:
          type: string
          format: date-time
      required: [id, owner_user_id, name, visibility, join_mode, created_at]

    JoinRequest:
      type: object
      properties:
        invite_token:
          type: string
          nullable: true

    InviteRequest:
      type: object
      properties:
        invited_user_id:
          type: string
          format: uuid
          nullable: true
        invited_email:
          type: string
          nullable: true
        expires_in_hours:
          type: integer
          minimum: 1
          maximum: 720
          default: 72

    InviteResponse:
      type: object
      properties:
        token:
          type: string
        channel_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
      required: [token, channel_id, expires_at]

    MembershipActionResponse:
      type: object
      properties:
        channel_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/MembershipRole'
      required: [channel_id, user_id, role]

    PublishMessageRequest:
      type: object
      description: Exactly one of `content_text` or `content_json` must be provided.
      properties:
        content_text:
          type: string
          nullable: true
        content_json:
          type: object
          nullable: true
          additionalProperties: true

    MessageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        channel_id:
          type: string
          format: uuid
        sender_user_id:
          type: string
          format: uuid
        seq_id:
          type: integer
          format: int64
        content_type:
          type: string
          enum: [text, json]
        content_text:
          type: string
          nullable: true
        content_json:
          type: object
          nullable: true
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - channel_id
        - sender_user_id
        - seq_id
        - content_type
        - created_at

    SeenRequest:
      type: object
      properties:
        last_seen_message_id:
          type: string
          format: uuid
          nullable: true
        last_seen_seq_id:
          type: integer
          format: int64
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true

    SeenResponse:
      type: object
      properties:
        channel_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        last_seen_seq_id:
          type: integer
          format: int64
          nullable: true
        unread_count:
          type: integer
      required: [channel_id, user_id, unread_count]